fix: rename namespaces
